@@ -26,11 +26,6 @@ use function in_array;
  */
 class WebInfo extends EntityRepository
 {
-    /**
-     * @param array|null $filter
-     *
-     * @return QueryBuilder
-     */
     public function findFiltered(array $filter = null): QueryBuilder
     {
         $queryBuilder = $this->_em->createQueryBuilder();
@@ -73,14 +68,6 @@ class WebInfo extends EntityRepository
         return $queryBuilder;
     }
 
-    /**
-     * SubSelect builder which limits the results of webInfos to only the active (Approved and FPP).
-     *
-     * @param QueryBuilder $queryBuilder
-     * @param array        $filter
-     *
-     * @return QueryBuilder
-     */
     public function applyWebInfoFilter(QueryBuilder $queryBuilder, array $filter): QueryBuilder
     {
         if (!empty($filter['search'])) {
@@ -94,7 +81,7 @@ class WebInfo extends EntityRepository
             );
 
 
-            $queryBuilder->setParameter('like', sprintf("%%%s%%", $filter['search']));
+            $queryBuilder->setParameter('like', sprintf('%%%s%%', $filter['search']));
         }
 
         return $queryBuilder;
